+++ src/components/Headline/index.tsx
@@ -16,8 +16,7 @@ export function Headline(props: Headline) {
     <div>
       <h1 className={classes.title}>{props.page} page</h1>
       <p className={classes.description}>
-        Get started by editing
-        {props.children}
+        アイテムの数は{props.children}個です。
       </p>
     </div>
   );
