+++ src/components/Main/index.tsx
@@ -1,18 +1,49 @@
 import { Headline } from "src/components/Headline";
 import classes from "src/components/Main/Main.module.css";
 import { Links } from "src/components/Links";
+import { useState, useCallback } from "react";
 
 type Main = {
   page: string;
 };
 
+const ITEMS = [
+  {
+    href: "https://nextjs.org/docs",
+    title: "Documentation \u2192",
+    description: "Find in-depth information about Next.js features and API.",
+  },
+  {
+    href: "https://nextjs.org/learn",
+    title: "Learn -> ",
+    description: "Learn about Next.js in an interactive course with quizzes!",
+  },
+  {
+    href: "https://github.com/vercel/next.js/tree/canary/examples",
+    title: "Examples -> ",
+    description: "Discover and deploy boilerplate example Next.js projects.",
+  },
+  {
+    href: "https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app",
+    title: "Deploy -> ",
+    description:
+      "Instantly deploy your Next.js site to a public URL with Vercel.",
+  },
+];
+
 export function Main(props: Main) {
+  const [items, setItems] = useState(ITEMS);
+  const handleReduce = useCallback(() => {
+    setItems((prevItems) => {
+      return prevItems.slice(0, prevItems.length - 1);
+    });
+  }, []);
   return (
     <main className={classes.main}>
       <Headline page={props.page}>
-        <code className={classes.code}>pages/{props.page}.tsx</code>
+        <code className={classes.code}>{items.length}</code>
       </Headline>
-      <Links />
+      <Links items={items} handleReduce={handleReduce} />
     </main>
   );
 }
